Created PlayerActions.cs
Created PlayerActions.cs and consolidated code from other scripts to make door open properly
