+++ FinalProj326/Assets/Scripts/PlayerActions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b3aca889b5e9d146857d3d7271b552d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
